+++ src/Components/MiniPlayer.js
@@ -53,94 +53,94 @@ export const MiniPlayer = ({ style }) => {
 
     if (!activeTrack) return;
     return (
-        <View
-            onPress={() => router.push("/player")}
-            style={[style, mainStyles.color_bg]}
-        >
-            <View style={mainStyles.miniPlayer}>
-                <View
-                    style={{
-                        flexDirection: "row",
-                        alignItems: "center",
-                        columnGap: 8,
-                    }}
-                >
-                    <AlbumArt
-                        image={activeTrack.artwork}
-                        style={{
-                            height: 48,
-                            aspectRatio: 1,
-                            borderRadius: 5,
-                        }}
-                    />
+        <View style={[style, mainStyles.color_bg]}>
+            <TouchableNativeFeedback
+                style={{ flex: 1 }}
+                onPress={() => router.push("/player")}
+            >
+                <View style={mainStyles.miniPlayer}>
                     <View
                         style={{
-                            flex: 1,
                             flexDirection: "row",
-                            justifyContent: "space-between",
                             alignItems: "center",
-                            width: "100%",
-                            columnGap: 16,
+                            columnGap: 8,
                         }}
                     >
+                        <AlbumArt
+                            image={activeTrack.artwork}
+                            style={{
+                                height: 48,
+                                aspectRatio: 1,
+                                borderRadius: 5,
+                            }}
+                        />
                         <View
                             style={{
                                 flex: 1,
-                                flexDirection: "column",
+                                flexDirection: "row",
+                                justifyContent: "space-between",
+                                alignItems: "center",
                                 width: "100%",
+                                columnGap: 16,
                             }}
                         >
-                            <Text style={textStyles.h4} numberOfLines={1}>
-                                {activeTrack.title}
-                            </Text>
-                            <Text style={textStyles.detail} numberOfLines={1}>
-                                {activeTrack.artist
-                                    ? activeTrack.artist
-                                    : "No artist"}
-                                {" • "}
-                                {activeTrack.date
-                                    ? new Date(activeTrack.date).getFullYear()
-                                    : "No date"}
-                            </Text>
-                        </View>
-                        {/* <View style={{ marginTop: -10, marginLeft: -13 }}>
-                            <IconButton
-                                onPress={handleLikeButtonPress}
-                                icon={
-                                    songData.isLiked ? "heart" : "heart-outline"
-                                }
-                                size={18}
-                            />
-                        </View> */}
-                        <View style={{ flexDirection: "row", columnGap: 16 }}>
-                            <IconButton
-                                onPress={previous}
-                                icon="skip-previous"
-                                size={36}
-                            />
-                            <IconButton
-                                onPress={hanldePlayPausePress}
-                                icon={
-                                    playbackState.state === "playing"
-                                        ? "pause"
-                                        : "play"
-                                }
-                                size={36}
-                            />
-                            <IconButton
-                                onPress={next}
-                                icon="skip-next"
-                                size={36}
-                            />
+                            <View
+                                style={{
+                                    flex: 1,
+                                    flexDirection: "column",
+                                    width: "100%",
+                                }}
+                            >
+                                <Text style={textStyles.h4} numberOfLines={1}>
+                                    {activeTrack.title}
+                                </Text>
+                                <Text
+                                    style={textStyles.detail}
+                                    numberOfLines={1}
+                                >
+                                    {activeTrack.artist
+                                        ? activeTrack.artist
+                                        : "No artist"}
+                                    {" • "}
+                                    {activeTrack.date
+                                        ? new Date(
+                                              activeTrack.date
+                                          ).getFullYear()
+                                        : "No date"}
+                                </Text>
+                            </View>
+                            <View
+                                style={{ flexDirection: "row", columnGap: 16 }}
+                            >
+                                <IconButton
+                                    onPress={previous}
+                                    icon="skip-previous"
+                                    size={36}
+                                />
+                                <IconButton
+                                    onPress={hanldePlayPausePress}
+                                    icon={
+                                        playbackState.state === "playing"
+                                            ? "pause"
+                                            : "play"
+                                    }
+                                    size={36}
+                                />
+                                <IconButton
+                                    onPress={next}
+                                    icon="skip-next"
+                                    size={36}
+                                />
+                            </View>
                         </View>
                     </View>
+                    <PlaybackSlider
+                        trackDuration={progress.duration}
+                        trackPosition={progress.position}
+                        skipPosition={seekToPosition}
+                    />
                 </View>
-                <PlaybackSlider
-                    trackDuration={progress.duration}
-                    trackPosition={progress.position}
-                    skipPosition={seekToPosition}
-                />
-            </View>
+            </TouchableNativeFeedback>
         </View>
     );
 };