+++ src/app/_layout.js
@@ -2,8 +2,11 @@ import { BottomSheetModalProvider } from "@gorhom/bottom-sheet";
 import { Stack } from "expo-router";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { colors } from "../styles/constants";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
 
 export default function App() {
+    const insets = useSafeAreaInsets();
+    
     return (
         <GestureHandlerRootView style={{ flex: 1, backgroundColor: colors.bg }}>
             <BottomSheetModalProvider>
@@ -27,8 +30,11 @@ function DefaultLayout() {
             <Stack.Screen
                 name="player"
                 options={{
-                    gestureDirection: "vertical",
+                    presentation: "modal",
                     animation: "slide_from_bottom",
+                    gestureEnabled: true,
+                    gestureDirection: "vertical",
+                    fullScreenGestureEnabled: true,
                 }}
             />
         </Stack>