+++ src/app/player.js
@@ -3,7 +3,7 @@ import React, { useEffect, useRef } from "react";
 import PlaybackControls from "../Components/PlaybackControls";
 import { mainStyles } from "../styles/styles";
 import ImageBlurBackground from "../Components/ImageBlurBackground";
-import { spacing } from "../styles/constants";
+import { colors, spacing } from "../styles/constants";
 import AlbumArt from "../Components/AlbumArt";
 import { useActiveTrack } from "react-native-track-player";
 import { textStyles } from "../styles/text";
@@ -11,6 +11,7 @@ import { useSongsStore } from "../store/songs";
 import PrimaryRoundIconButton from "../Components/UI/PrimaryRoundIconButton";
 import SongSheet from "../Components/BottomSheets/SongSheet";
 import IconButton from "../Components/UI/IconButton";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
 
 const PlayerScreen = () => {
     const { addSongLike, removeSongLike, getSong } = useSongsStore();
@@ -33,14 +34,20 @@ const PlayerScreen = () => {
         }
     };
 
+    const insets = useSafeAreaInsets();
     return (
         <View style={[mainStyles.container, { justifyContent: "center" }]}>
-            <ImageBlurBackground />
+            <ImageBlurBackground
+                image={activeTrack?.artwork}
+                styles={{ height: "85%", top: 0 }}
+            />
+            <DismissPill insets={insets} />
             <View
                 style={{
                     flex: 1,
                     justifyContent: "center",
                     marginHorizontal: spacing.appPadding,
+                    rowGap: spacing.sm,
                 }}
             >
                 <View>
@@ -96,6 +103,29 @@ const PlayerScreen = () => {
                 <PlaybackControls />
             </View>
             <SongSheet ref={editSongRef} />
+        </View>
+    );
+};
+
+const DismissPill = ({ insets = 0 }) => {
+    return (
+        <View
+            style={{
+                position: "absolute",
+                top: insets.top + spacing.sm,
+                right: 0,
+                left: 0,
+                alignItems: "center",
+            }}
+        >
+            <View
+                style={{
+                    width: 50,
+                    height: 7,
+                    backgroundColor: colors.primary,
+                    borderRadius: spacing.round,
+                }}
+            />
         </View>
     );
 };