+++ src/store/songs.js
@@ -1,14 +1,15 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { Audio, InterruptionModeAndroid } from "expo-av";
 import { create } from "zustand";
 import { createJSONStorage, persist } from "zustand/middleware";
 
+import { router } from "expo-router";
 import TrackPlayer, {
     AppKilledPlaybackBehavior,
     Capability,
+    RatingType,
     RepeatMode,
 } from "react-native-track-player";
-import { router } from "expo-router";
+import { PlaybackService } from "../../PlaybackService";
 
 const MARKED_SONGS_KEY = "MarkedSongs";
 
@@ -21,61 +22,59 @@ export const useSongsStore = create(
                     id: "1",
                     name: "Liked songs",
                     description: "Your songs that you liked.",
-                    image: null,
+                    artwork: null,
                     year: null,
                     artist: null,
                     date: null,
                     songs: [], //contains only id's
                 },
             ],
+            albums: [],
 
             // for menus, ect..
             selectedSong: null,
             selectedPlaylist: null,
-            activeBottomSheet: null, // would be cool maybe?
+            activeBottomSheets: [], // would be cool maybe?
+            safeAreaInsets: 0,
+            setSafeAreaInsets: (insets) => {
+                set({ safeAreaInsets: insets });
+            },
 
             // for music playback
-            currentTrack: null, //song id
-            playlist: [],
-            audioRef: null,
-            isPlaying: false,
             repeat: false,
-
-            trackDuration: 0,
-            trackPosition: 0,
-
             playingFrom: null,
-            isSetup: false,
-            playbackState: null, // would be nice for me to use this variable as a check in other components to change the ui
-            setPlaybackState: (state) => set({ playbackState: state }),
+
+            setRepeat: (mode) => {},
+            shuffle: () => {},
 
             setup: async () => {
-                if (get().isSetup) return console.log("already setup");
+                try {
+                    await TrackPlayer.getActiveTrack();
+                } catch (error) {
+                    TrackPlayer.registerPlaybackService(() => PlaybackService);
 
-                await TrackPlayer.setupPlayer({
-                    autoHandleInterruptions: true,
-                });
-                await TrackPlayer.updateOptions({
-                    progressUpdateInterval: 500,
-                    android: {
-                        appKilledPlaybackBehavior:
-                            AppKilledPlaybackBehavior.StopPlaybackAndRemoveNotification,
-                    },
-                    capabilities: [
-                        Capability.Play,
-                        Capability.Pause,
-                        Capability.SkipToNext,
-                        Capability.SkipToPrevious,
-                        Capability.SeekTo,
-                    ],
-                    compactCapabilities: [
-                        Capability.Play,
-                        Capability.Pause,
-                        Capability.SkipToNext,
-                    ],
-                });
-                await TrackPlayer.setRepeatMode(RepeatMode.Off);
-                set({ isSetup: true });
+                    await TrackPlayer.setupPlayer({
+                        autoHandleInterruptions: true,
+                    });
+
+                    await TrackPlayer.updateOptions({
+                        progressUpdateInterval: 250,
+                        ratingType: RatingType.Heart,
+                        android: {
+                            appKilledPlaybackBehavior:
+                                AppKilledPlaybackBehavior.StopPlaybackAndRemoveNotification,
+                        },
+                        notificationCapabilities: [
+                            Capability.Play,
+                            Capability.Pause,
+                            Capability.SkipToNext,
+                            Capability.SkipToPrevious,
+                            Capability.SeekTo,
+                            Capability.SetRating,
+                        ],
+                    });
+                    await TrackPlayer.setRepeatMode(RepeatMode.Queue);
+                }
             },
             addToQueue: async (song) => {
                 await TrackPlayer.add(song);
@@ -131,7 +130,7 @@ export const useSongsStore = create(
                             id: "1",
                             name: "Liked songs",
                             description: "Your songs that you liked.",
-                            image: null,
+                            artwork: null,
                             year: null,
                             artist: null,
                             date: null,
@@ -141,186 +140,6 @@ export const useSongsStore = create(
                 });
             },
 
-            // Init & cleanup for playback
-            loadTrack: async (song, playlist = null, shuffle = false) => {
-                if (!song) return console.log("No song provided.");
-
-                await get().unloadTrack();
-                try {
-                    Audio.setAudioModeAsync({
-                        staysActiveInBackground: true,
-                        playsInSilentModeIOS: true,
-                        interruptionModeAndroid:
-                            InterruptionModeAndroid.DoNotMix,
-                    });
-                    const { sound } = await Audio.Sound.createAsync(
-                        { uri: song.uri },
-                        { shouldPlay: true }
-                    );
-
-                    sound._onPlaybackStatusUpdate =
-                        get().setOnPlaybackStatusUpdate;
-
-                    set({
-                        audioRef: sound,
-                        currentTrack: song.id,
-                        playlist: playlist ? playlist : get().playlist,
-                        isPlaying: true,
-                        songs: get().songs.map((s) =>
-                            s.id === song.id
-                                ? {
-                                      ...s,
-                                      timesPlayed: s.timesPlayed + 1,
-                                      lastPlayed: new Date(),
-                                  }
-                                : s
-                        ),
-                    });
-                    if (shuffle) get().shuffle();
-                } catch (error) {
-                    console.error("Error loading audio:", error);
-                }
-            },
-
-            setOnPlaybackStatusUpdate: (playbackStatus) => {
-                if (playbackStatus.isLoaded) {
-                    set({
-                        trackPosition: playbackStatus.positionMillis,
-                        trackDuration: playbackStatus.durationMillis,
-                    });
-
-                    if (
-                        playbackStatus.didJustFinish &&
-                        !playbackStatus.isLooping
-                    ) {
-                        get().next();
-                    }
-                }
-            },
-
-            unloadTrack: async () => {
-                const audioRef = get().audioRef;
-                if (audioRef) {
-                    await audioRef.unloadAsync();
-                    set({ audioRef: null, isPlaying: false });
-                }
-            },
-
-            // Playback & controls
-            // play: async () => {
-            //     const audioRef = get().audioRef;
-            //     if (!audioRef) return;
-
-            //     try {
-            //         await audioRef.playAsync();
-            //         set({ isPlaying: true });
-            //     } catch (error) {
-            //         console.error("Error playing audio:", error);
-            //     }
-            // },
-            // pause: async () => {
-            //     const audioRef = get().audioRef;
-            //     if (!audioRef) return;
-
-            //     try {
-            //         await audioRef.pauseAsync();
-            //         set({ isPlaying: false });
-            //     } catch (error) {
-            //         console.error("Error pausing audio:", error);
-            //     }
-            // },
-
-            // next: async () => {
-            //     const audioRef = get().audioRef;
-            //     if (!audioRef) return;
-
-            //     const playlist = get().playlist.songs;
-            //     const currentTrack = get().currentTrack;
-
-            //     const index = playlist.indexOf(currentTrack);
-            //     const nextSongs = playlist.slice(index + 1);
-            //     const nextSong = nextSongs.find(
-            //         (songId) => !get().getSong(songId).isHidden
-            //     );
-
-            //     if (nextSong) {
-            //         const nextSongData = get().getSong(nextSong);
-            //         await get().loadTrack(nextSongData);
-            //     } else {
-            //         await get().unloadTrack();
-            //     }
-            // },
-
-            // previous: async () => {
-            //     const audioRef = get().audioRef;
-            //     if (!audioRef) return;
-
-            //     const playlist = get().playlist.songs;
-            //     const currentTrack = get().currentTrack;
-            //     const currentSong = get().getSong(currentTrack);
-
-            //     const index = playlist.indexOf(currentTrack);
-            //     const previousSongs = playlist.slice(0, index).reverse();
-            //     const previousSong = previousSongs.find(
-            //         (songId) => !get().getSong(songId).isHidden
-            //     );
-
-            //     if (previousSong) {
-            //         const previousSongData = get().getSong(previousSong);
-            //         const { positionMillis } = await audioRef.getStatusAsync();
-            //         if (positionMillis >= 1000) {
-            //             await audioRef.setPositionAsync(0);
-            //         } else {
-            //             await get().loadTrack(previousSongData);
-            //         }
-            //     }
-            // },
-
-            skipPosition: async (position) => {
-                const audioRef = get().audioRef;
-                if (!audioRef) return;
-                const { durationMillis } = await audioRef.getStatusAsync();
-                const newPosition = Math.floor(position * durationMillis);
-                await audioRef.setPositionAsync(newPosition);
-            },
-
-            shuffle: () => {
-                const playlist = get().playlist;
-
-                const currentSongId = get().currentTrack;
-                if (playlist.length < 2) return;
-
-                const shuffledSongs = [...playlist.songs];
-                const currentSongIndex = shuffledSongs.indexOf(currentSongId);
-                shuffledSongs.splice(currentSongIndex, 1);
-
-                for (let i = shuffledSongs.length - 1; i > 0; i--) {
-                    const j = Math.floor(Math.random() * (i + 1));
-                    [shuffledSongs[i], shuffledSongs[j]] = [
-                        shuffledSongs[j],
-                        shuffledSongs[i],
-                    ];
-                }
-                shuffledSongs.unshift(currentSongId);
-                set({ playlist: { ...playlist, songs: shuffledSongs } });
-            },
-
-            turnOnRepeat: () => {
-                const audioRef = get().audioRef;
-                if (!audioRef) return;
-
-                audioRef.setIsLoopingAsync(true);
-                set({ repeat: true });
-            },
-
-            turnOffRepeat: () => {
-                const audioRef = get().audioRef;
-                if (!audioRef) return;
-
-                audioRef.setIsLoopingAsync(false);
-                set({ repeat: false });
-            },
-
             // songs, selectedSong
             setSongs: (songs) => set({ songs }),
             getSong: (id) => get().songs.find((song) => song.id === id),
@@ -328,36 +147,18 @@ export const useSongsStore = create(
 
             inheritPlatlistDataToSongs: (playlistId) => {
                 const playlist = get().getPlaylist(playlistId);
-                const image = playlist.image;
+                const artwork = playlist.artwork;
                 const artist = playlist.artist;
                 const date = playlist.date;
 
                 set((state) => ({
                     songs: state.songs.map((song) =>
                         playlist.songs.includes(song.id)
-                            ? { ...song, image, artist, date }
+                            ? { ...song, artwork, artist, date }
                             : song
                     ),
                 }));
             },
-
-            // deleteSongFromDevice: async (id) => {
-            //     set((state) => ({
-            //         songs: state.songs.filter((song) => song.id !== id),
-            //     }));
-
-            //     const songUri = get().getSong(id).uri;
-            //     const playlists = get().playlists;
-            //     set({
-            //         playlists: playlists.map((playlist) => ({
-            //             ...playlist,
-            //             songs: playlist.songs.filter((songId) => songId !== id),
-            //         })),
-            //     });
-            //     await FileSystem.deleteAsync(songUri).catch((e) =>
-            //         console.log("error deleting song", e)
-            //     );
-            // },
 
             // like, unlike
             addSongLike: (id) => {
@@ -394,7 +195,13 @@ export const useSongsStore = create(
             setSelectedPlaylist: (playlist) =>
                 set({ selectedPlaylist: playlist }),
 
-            createPlaylist: (name, description, image = null, artist, date) => {
+            createPlaylist: (
+                name,
+                description,
+                artwork = null,
+                artist,
+                date
+            ) => {
                 name = name ? name : "Unnamed playlist";
                 const newPlaylist = {
                     id: (
@@ -403,7 +210,7 @@ export const useSongsStore = create(
                     ).toUpperCase(),
                     name,
                     description,
-                    image,
+                    artwork,
                     artist,
                     date,
                     songs: [],
@@ -413,7 +220,7 @@ export const useSongsStore = create(
                 }));
             },
 
-            editPlaylist: (id, name, description, image, artist, date) => {
+            editPlaylist: (id, name, description, artwork, artist, date) => {
                 name = name ? name : "Unnamed playlist";
                 set((state) => ({
                     playlists: state.playlists.map((playlist) =>
@@ -424,7 +231,7 @@ export const useSongsStore = create(
                                   ...(description !== ""
                                       ? { description }
                                       : {}),
-                                  ...(image !== null ? { image } : {}),
+                                  ...(artwork !== null ? { artwork } : {}),
                                   ...(artist !== null ? { artist } : {}),
                                   ...date,
                               }
@@ -520,77 +327,6 @@ export const useSongsStore = create(
                     }));
                 }
             },
-        }),
-        {
-            name: MARKED_SONGS_KEY,
-            storage: createJSONStorage(() => AsyncStorage),
-            partialize: (state) => ({
-                songs: state.songs,
-                playlists: state.playlists,
-            }),
-        }
-    )
-);
-
-export const useAudioStore = create(
-    persist(
-        (set, get) => ({
-            songs: [
-                {
-                    id: 1,
-                    uri: null, // file path
-                    duration: 0, // in milliseconds
-
-                    name: "song3",
-                    image: null, // image path
-                    artist: null, // artist name
-                    albumId: null, // album id
-
-                    liked: false,
-                    hidden: false,
-                    statistics: {
-                        lastPlayed: new Date(), // date when song was last played
-                        timesPlayed: 0,
-                        timesSkipped: 0,
-                    },
-                },
-            ],
-            playlists: [
-                {
-                    id: 1,
-                    name: "playlist2",
-                    description: "Your songs that you liked.",
-                    image: null, // image path
-                    created_at: new Date(),
-                    songs: [], // array of song id's
-                },
-            ],
-            albums: [
-                {
-                    id: 1,
-                    name: "album1",
-                    artist: null, // artist name
-                    date: new Date(), // date album was created
-                },
-            ],
-            statistics: {
-                totalPlaytime: 0, // in seconds
-                history: [], // array of song id's
-            },
-
-            // for menus, ect..
-            selectedSong: null,
-            selectedPlaylist: null,
-
-            // for music playback
-            currentTrack: null, //song id that is currently playing
-            queue: [], // array of song id's
-            audioRef: null,
-            isPlaying: false,
-            onRepeat: false,
-
-            currentTrackDuration: 0,
-            currentTrackPosition: 0,
         }),
         {
             name: MARKED_SONGS_KEY,